h5website-backend/requirements.txt
pandas
chromadb
fastapi
uvicorn
tiktoken
google-generativeai